Improve consistency of final state of coreos prometheus
Avoid downloading to the final target directory, using move and clean up
of previous attempts to ensure a final consistent result.

Ensure that files added via listdir are watched which is only enabled
when recursive is enabled.

Move the creation of the cache dir to the set up and retrieval of the
cache path itself and allow for parent components not to exist.