test-in-tilt: fix for Go 1.16
These were running module-mode incompliant commands - when you
have a `go.mod`, you MUST run the `go test` command from within
the module hierarchy (i.e. at or under the `go.mod` dir).

In practice, this isn't a huge issue, as generally you'd have a
`Tiltfile` at the same level as a `go.mod`, so the test structure
is adjusted to reflect this with the elimination of the `fake_project`
dir.

It is possible to work around this using the new `dir` argument
that `local_resource` supports, which will eventually receive it
via `**kwargs`, so if you have the file structure:

```
├── go-app
│   ├── go.mod
│   └──main.go
└── Tiltfile
```

You can do:
```
test_go('mytests', './...', './go-app', dir='./go-app')
```

So that the `go test ./...` command is run inside of the `go-app`
directory, instead of trying to pass `./go-app/...` as the "package".