ko: fix tests for Go 1.16
The package we were building does not build in Go module-aware
mode. It was also a random third-party package.

This replaces it with a tiny hello world local app. While this
does not test the remote fetch functionality of the extension,
that is actually largely inoperable with modern projects (i.e.
those that use Go modules).

See https://github.com/tilt-dev/tilt-extensions/issues/188 for
details.