ci: fixes for Go 1.16 (#186)
* ci: use pinned version of image

Avoid large breakage in the future by making it easier to test new
versions of the CI image pre-merge.

* test-in-tilt: fix for Go 1.16

These were running module-mode incompliant commands - when you
have a `go.mod`, you MUST run the `go test` command from within
the module hierarchy (i.e. at or under the `go.mod` dir).

In practice, this isn't a huge issue, as generally you'd have a
`Tiltfile` at the same level as a `go.mod`, so the test structure
is adjusted to reflect this with the elimination of the `fake_project`
dir.

It is possible to work around this using the new `dir` argument
that `local_resource` supports, which will eventually receive it
via `**kwargs`, so if you have the file structure:

```
├── go-app
│   ├── go.mod
│   └──main.go
└── Tiltfile
```

You can do:
```
test_go('mytests', './...', './go-app', dir='./go-app')
```

So that the `go test ./...` command is run inside of the `go-app`
directory, instead of trying to pass `./go-app/...` as the "package".

* ko: fix tests for Go 1.16

The package we were building does not build in Go module-aware
mode. It was also a random third-party package.

This replaces it with a tiny hello world local app. While this
does not test the remote fetch functionality of the extension,
that is actually largely inoperable with modern projects (i.e.
those that use Go modules).

See https://github.com/tilt-dev/tilt-extensions/issues/188 for
details.