Update coreos prometheus deployment to release v0.8.0
Roll forward to the 0.8.0 release of coreos kube-prometheus which will
pull in a more recent grafana dashboard.

Requires patching the node exporter deployment in addition to the
previous adjustments. This is due to when running in kind or k3d
environments cannot mount /sys from the host directly and instead switch
back to settings used in previous releases.