kubefwd/ngrok: use portable mktemp invocation for macOS/Linux (#312)
BSD `mktemp` has some differences compared to the GNU variant,
so this is the most portable way to make temporary files that
works across macOS + Linux that I could find. Note that the ngrok
file won't end in `.yml` anymore because BSD `mktemp` only allows
the randomized part of the template (i.e. `XXXXXX`) to be at the
END of the filename.

Both of these assume a POSIX-y env as-is (e.g. they're running
Bash scripts for their main operator functionality, so no attempt
at supporting Windows cmd is made here.