uibutton: allow nav buttons to log to resource
Sometimes it's convenient to have a button in the navbar but
still want its output to go to a specific resource.

In the default case, we also now choose the `(Tiltfile)` resource
instead of sending it to the all resources/global log view, which
can be hard to find the output in.

See tilt-dev/tilt#5381.