helm_resource: only modify metadata
Extract metadata: and trailing indented lines, and only perform namespace
insertion on that text. This was causing some charts to fail to deploy, for
example if a chart has a CRD with an openapi spec that had a `namespace:` key,
the key was erroneously removed by the previous apply helper.