coreos_prometheus: bump to 0.15, remove operator dep (#639)
This bumps the kube-prometheus operator release to 0.15.0. This is
because according to this [1] compatibility table, 0.14.0 is not
supported by versions Kubernetes >= 1.32

This also removes the operator resource dependency for the CRDs. This
inclusion means that the operator pod is started before the CRDs. This
results in the pod not being able to read Prometheus or Alertmanager CRD
resources, so the resulting pods are never created until the operator
pod is restarted. Creating the CRDs before the operator pod fixes this
problem.

Signed-off-by: Ben Kero <bkero@bke.ro>